+++ src/Project/BuildSteps/Preload.ts
@@ -9,6 +9,24 @@ import preload from 'openui5-preload';
 import Ui5 from '../../Utils/Ui5Vscode';
 
 export default {
+  async clean(ui5Project: Ui5Project, destPath: string): Promise<void> {
+    const { isLibrary } = ui5Project;
+
+    const aFiles = isLibrary ? [`library-preload.json`, `library-preload.js`] : [`Component-preload.js`];
+    for (let i = 0; i < aFiles.length; i++) {
+      const uriToDelete = Uri.file(path.join(destPath, aFiles[i]));
+      Log.builder(`Deleting file ${uriToDelete.fsPath}`);
+      try {
+        await workspace.fs.delete(uriToDelete, {
+          useTrash: false,
+        });
+        Log.builder(`Deleted successfully`);
+      } catch (oError) {
+        Log.builder(`File does not exist`);
+      }
+    }
+  },
+
   async build(ui5Project: Ui5Project, srcPath: string, destPath: string): Promise<void> {
     const { isLibrary, namespace } = ui5Project;